Extracted useRegistry hook into its own file
