+++ src/contexts/index.tsx
@@ -2,3 +2,4 @@ export { SessionProvider, SessionProviderContext } from './SessionProvider'
 export { ApiProvider, ApiProviderContext } from './ApiProvider'
 export { ThemeProvider, ThemeProviderContext } from './ThemeProvider'
 export { WebSocketProvider, WebSocketProviderContext } from './WebSocketProvider'
+export { RegistryProvider, RegistryContext } from './RegistryProvider'
