+++ src/hooks/index.tsx
@@ -5,3 +5,4 @@ export { useResize } from '@/navigation/hooks/useResize'
 export { useSession } from './useSession'
 export { useQuery } from './useQuery'
 export { useApi } from './useApi'
+export { useRegistry } from './useRegistry'
